+++ crates/client/src/serialize/txt/rdata_parsers/openpgpkey.rs
@@ -23,9 +23,9 @@ use crate::rr::rdata::OPENPGPKEY;
 ///    of [RFC4648].
 /// ```
 pub fn parse<'i, I: Iterator<Item = &'i str>>(mut tokens: I) -> ParseResult<OPENPGPKEY> {
-    let encoded_public_key = tokens
-        .next()
-        .ok_or_else(|| ParseErrorKind::Message("OPENPGPKEY public key field is missing"))?;
+    let encoded_public_key = tokens.next().ok_or(ParseErrorKind::Message(
+        "OPENPGPKEY public key field is missing",
+    ))?;
     let public_key = data_encoding::BASE64.decode(encoded_public_key.as_bytes())?;
     Some(OPENPGPKEY::new(public_key))
         .filter(|_| tokens.next().is_none())