+++ crates/client/src/serialize/txt/rdata_parsers/tlsa.rs
@@ -49,12 +49,12 @@ pub fn parse<'i, I: Iterator<Item = &'i str>>(tokens: I) -> ParseResult<TLSA> {
 
     let token = iter
         .next()
-        .ok_or_else(|| ParseErrorKind::Message("TLSA selector field missing"))?;
+        .ok_or(ParseErrorKind::Message("TLSA selector field missing"))?;
     let selector = to_u8(token)?.into();
 
     let token = iter
         .next()
-        .ok_or_else(|| ParseErrorKind::Message("TLSA matching field missing"))?;
+        .ok_or(ParseErrorKind::Message("TLSA matching field missing"))?;
     let matching = to_u8(token)?.into();
 
     // these are all in hex: "a string of hexadecimal characters"