+++ crates/client/src/serialize/txt/zone.rs
@@ -412,7 +412,7 @@ impl Parser {
                 // TODO, should these all be checked operations?
                 '0'..='9' => {
                     collect *= 10;
-                    collect += c.to_digit(10).ok_or_else(|| ParseErrorKind::CharToInt(c))?;
+                    collect += c.to_digit(10).ok_or(ParseErrorKind::CharToInt(c))?;
                 }
                 'S' | 's' => {
                     value += collect;